Switch the generic argument order for Coro: <Y, R, I> -> <I, Y, R>.
I find this order more intuitive, at the expense of not being able
to have a default value for the parameter (I) that has the most
sane default. But in practice I've found that all of the example
code explicitly specifies the I parameter anyway, so the benefit
of ordering it this way (I before Y or R) seems greater than the
benefit of allowing elision when I = ().