Make Void uninhabited and update documentation.
The way Void was rendered uninstantiable before was by making it
wrap a hidden private type and left it to the honor system for
nothing to construct that private type elsewhere in the module.
But there's a better way: an empty enum. This is how the Infallible
type in the standard library is implemented, and enums implemented
this way have special support as of Rust 1.82, where you can write
`match` statements and omit branches for uninhabited types.

This makes it possible to match the suspended state of a `Coro`
whose return type is `Void` and omit the branch to handle the
`Return` variant, making the indirection from `to_enum()` less
necessary. However, it's still possible that using
`SuspendedVisitor` instead of `match` in performance-critical areas
could attain a performance improvement since it's not clear how the
compiler represents an enum where all but one of the variants are
uninhabited; is it smart enough to omit the discriminator? More
research is needed.