Make compose() return a concrete struct, not an impl trait.
This allows the compiler to conserve derived traits like Clone and
FixedPointCoro that can be inferred from the properties of the
Compose<> parameters. Unfortunately `impl Trait` erases a lot of
information that might be needed, a bitter lesson for API design.

We desperately need some kind of syntax where we can express
derived properties on `impl Trait` return types that depend on the
properties of input types, something like:

```
fn compose<Y2, K2>(
    self,
    other: K2,
) -> impl Coro<I, Y2, R>
    + Clone if { Self: Clone, K2: Clone }
    + Copy if { Self: Copy, K2: Copy }
    + FixedPointCoro<I, Y2, R> if {
        Self: FixedPointCoro<I, Y, R>,
        K2: FixedPointCoro<Y, Y2, R>,
    }
where
    K2: Coro<Y, Y2, R>,
{
    Compose::new(self, other)
}
```

Note to self to bikeshed a trait algebra RFC but in the meantime
stick to returning concrete types.