Add a fixed_point() no-op method to FixedPointCoro.
If you have a complicated pipeline of coroutines and combinators
that you expect to ultimately implement `FixedPointCoro`, but find
that it's breaking down somewhere, it can be helpful to insert
calls to `.fixed_point()` in the pipeline to debug. The compiler
will show an error on the first call to `.fixed_point()` on a
coroutine expression that isn't `FixedPointCoro`.