Add absurd() method to Void.
This can be used to implement generic traits where the fact that
Void is an input parameter guarantees that the code is unreachable.

One can simply use the unreachable!() macro directly but this
better documents that it's the fact that a Void is in scope that
makes the code unreachable. One can audit their codebase for uses
of the unreachable!() macro (or similar) and rely on the type
safety of Void to hold the guarantee the invariants that the
unreachable!() macro cannot on its own.