Change take() from a combinator to a coroutine.
This is slightly more verbose but more elegant overall; it allows
the logic underlying take() to be used in more contexts, and it
demonstrates the power of compose(). Furthermore, it reduces how
often one needs to use returns::<Void>() when working with
infinite coroutines made finite by composing with take(), and
lets those infinite coroutines be agnostic about their return
type and generally avoid using Void. This also gives the
programmer more flexibility in how to assign return values for
the case where the source returns only or the take() coroutine
reaches its limit, using map_return() primitives as shown by the
examples.