Add new drive() method and refactor for_each() based on it.
This is a more general version of for_each() that works with
input types that don't implement Default, and allows the user to
control what inputs the coroutine will be resumed with at the
next step for each yielded value.

The for_each() function is rewritten in terms of drive().